Code adjustments. Int types modified

+++ array_of_products.c
@@ -3,21 +3,21 @@
 #define LIST_SIZE 5
 
 static void
-printResult_(int32_t *outputList, int32_t listSize)
+printResult_(int64_t *outputList, int64_t listSize)
 {
     if (NULL != outputList)
     {
-        for (int32_t index = 0; index < listSize; index++)
+        for (int64_t index = 0; index < listSize; index++)
         {
-            printf("outputList[%d] = %d\n", index, *(outputList + index));
+            printf("outputList[%ld] = %ld\n", index, *(outputList + index));
         }
     }
 }
 
 int main()
 {
-    int32_t listOfNumbers[LIST_SIZE] = {1, 2, 3, 4, 5};
-    int32_t productList[LIST_SIZE];
+    int64_t listOfNumbers[LIST_SIZE] = {1, 2, 3, 4, 5};
+    int64_t productList[LIST_SIZE];
 
     vsBool rc = setProducList(listOfNumbers,
                               LIST_SIZE,
@@ -33,17 +33,19 @@ int main()
 }
 
 static void
-arrayOfProducts_(const int32_t *list,
-                 int32_t currentIndex,
-                 int32_t *listOutput,
-                 const int32_t *finishBound)
+arrayOfProducts_(const int64_t *list,
+                 int64_t currentIndex,
+                 int64_t *listOutput,
+                 const int64_t *finishBound)
 {
-    int32_t product = 1;
+    // printf("Function call: arrayOfProducts_\n");
+    int64_t product = 1;
 
-    for (int32_t index = 0;; index++)
+    for (int64_t index = 0;; index++)
     {
         if ((list + index) != (list + currentIndex))
         {
+            // printf("calculating product...\n");
             product *= *(list + index);
         }
 
@@ -53,6 +55,7 @@ arrayOfProducts_(const int32_t *list,
         }
     }
 
+    // printf("added to outputList!\n\n");
     listOutput[currentIndex] = product;
 
     if ((list + currentIndex) != finishBound)
@@ -62,12 +65,12 @@ arrayOfProducts_(const int32_t *list,
 }
 
 vsBool
-setProducList(int32_t *numbersList,
-              int32_t numbersListSize,
-              int32_t *outputList,
-              int32_t outputListSize)
+setProducList(int64_t *numbersList,
+              int64_t numbersListSize,
+              int64_t *outputList,
+              int64_t outputListSize)
 {
-    static int32_t *finishBound_ = NULL;
+    static int64_t *finishBound_ = NULL;
 
     vsBool rc = TRUE;
